Refactored for modularity and reduced repetition
